refactor: clean components
