+++ src/client/components/_commons/ChatUI/AssistantMessageText.tsx
@@ -1,0 +1,12 @@
+import { marked } from "marked";
+
+export const AssistantMessageText = ({ text }: { text: string }) => {
+  return (
+    <div
+      className="albert-markdown"
+      dangerouslySetInnerHTML={{
+        __html: marked.parse(text, { async: false }) as string,
+      }}
+    />
+  );
+};
